Fixed the __gt__ and __lt__ operators on attributes
By using the newly implemented inclusive/exclusive support in the Range class
from the Filters API, this commit fixes the issue #8 from the tracker.

Although, there's some weird bug on the Filters API that is messing with this
fix somehow. Apparently, the init method on the FilterProcessor is calling
the Range constructor and ignoring the leftInclusive, rightInclusive
attributes from the object, which should be passed along to the newly created
objects as appropriate.