Updating plugin for gephi 0.9
It compiles, and display a window in Gephi. But:
(i) the import of data does not work : an exception occurs that prevent
the graph creation, thus the nodes and edges keep empty;
(ii) the SOAP client library was removed due to issues on the wsdl
files: it seems it should be rewritten following the format followed by
the current tools.